Execute tasks for included builds in a separate thread
IncludedBuildController allows tasks to be queued for execution,
and once started will loop and attempt to execute all queued
tasks. This execution loop is performed in a dedicated thread,
with the threads from dependent tasks blocking until the
required task is completed.

This mechanism will permit included build tasks to be executed
in parallel with other included builds, and will also naturally
lead to a separate build operation hierarchy for included build
tasks.