Added an initial 'cpp-executable' plugin. This plugin compiles and links C++ source files found in `src/main/cpp` and headers found in `src/main/headers` into a native executable.
Builds only for the current platform using a hard-coded selection of tool chains. Does not support any dependencies, or any configurability beyond that offered by the `CppCompile` and `LinkExecutable` task types.

This plugin lives in the 'current' model and is intended to behave very similarly to the 'java-application' plugin. It, along with some further new plugins, is intended to eventually replace the software model native plugin.